+++ src/AutoRest.CSharp/Common/AutoRest/Plugins/Configuration.cs
@@ -36,9 +36,9 @@ namespace AutoRest.CSharp.Input
             public const string SuppressAbstractBaseClasses = "suppress-abstract-base-class";
 
             // options added for compatibility with v2 generators
+            public const string CompatClientFactory = "compat-client-factory";
             public const string CompatClientInterfaces = "compat-client-interfaces";
             public const string CompatErrorResponses = "compat-error-responses";
-            public const string CompatAnonymousClientCtor = "compat-anonymous-client-ctor";
         }
 
         public static void Initialize(
@@ -59,9 +59,9 @@ namespace AutoRest.CSharp.Input
             string? projectFolder,
             string[] protocolMethodList,
             IReadOnlyList<string> suppressAbstractBaseClasses,
+            bool compatClientFactory,
             bool compatClientInterfaces,
             bool compatErrorResponses,
-            bool compatAnonymousClientCtor,
             MgmtConfiguration mgmtConfiguration)
         {
             _outputFolder = outputFolder;
@@ -78,9 +78,9 @@ namespace AutoRest.CSharp.Input
             SingleTopLevelClient = singleTopLevelClient;
 
             // compat options
+            CompatClientFactory = compatClientFactory;
             CompatClientInterfaces = compatClientInterfaces;
             CompatErrorResponses = compatErrorResponses;
-            CompatAnonymousClientCtor = compatAnonymousClientCtor;
 
             projectFolder ??= ProjectFolderDefault;
             if (Path.IsPathRooted(projectFolder))
@@ -117,9 +117,9 @@ namespace AutoRest.CSharp.Input
         public static bool SingleTopLevelClient { get; private set; }
         public static bool SkipSerializationFormatXml { get; private set; }
         public static bool DisablePaginationTopRenaming { get; private set; }
+        public static bool CompatClientFactory { get; private set; }
         public static bool CompatClientInterfaces { get; private set; }
         public static bool CompatErrorResponses { get; private set; }
-        public static bool CompatAnonymousClientCtor { get; private set; }
 
         private static IReadOnlyList<string>? _suppressAbstractBaseClasses;
         public static IReadOnlyList<string> SuppressAbstractBaseClasses => _suppressAbstractBaseClasses ?? throw new InvalidOperationException("Configuration has not been initialized");
@@ -155,9 +155,9 @@ namespace AutoRest.CSharp.Input
                 projectFolder: autoRest.GetValue<string?>(Options.ProjectFolder).GetAwaiter().GetResult(),
                 protocolMethodList: autoRest.GetValue<string[]?>(Options.ProtocolMethodList).GetAwaiter().GetResult() ?? Array.Empty<string>(),
                 suppressAbstractBaseClasses: autoRest.GetValue<string[]?>(Options.SuppressAbstractBaseClasses).GetAwaiter().GetResult() ?? Array.Empty<string>(),
+                compatClientFactory: GetOptionValue(autoRest, Options.CompatClientFactory),
                 compatClientInterfaces: GetOptionValue(autoRest, Options.CompatClientInterfaces),
                 compatErrorResponses: GetOptionValue(autoRest, Options.CompatErrorResponses),
-                compatAnonymousClientCtor: GetOptionValue(autoRest, Options.CompatAnonymousClientCtor),
                 mgmtConfiguration: MgmtConfiguration.GetConfiguration(autoRest)
             );
         }
@@ -191,11 +191,11 @@ namespace AutoRest.CSharp.Input
                     return false;
                 case Options.DisablePaginationTopRenaming:
                     return false;
+                case Options.CompatClientFactory:
+                    return false;
                 case Options.CompatClientInterfaces:
                     return false;
                 case Options.CompatErrorResponses:
-                    return false;
-                case Options.CompatAnonymousClientCtor:
                     return false;
                 default:
                     return null;